+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Improved the flexibility of the WhoisServerList tests.
 - Refactor AbstractClient and client to use the AbstractLocator and DomainLocator.
 - Update AbstractClient property name - from tldLocator to whoisLocator  - to better reflect usage.
-
+- Refactor tests to use DomainLocator instead of removed Locator class.
 
 ### Added
 - Create an AbstractLocator and DomainLocator for more flexibility!